Drush install issues
Both DE installs had problems with composer not automatically creating a composer.json file in the home directory.
When I had went to use drush, a lot of the commands were not working. Not to mention Composer wouldnt update drush to the newest version.
This fixed my issue on both installations.